allow rosdep rule to specify explicit python package
This would allow rosdep rules like this:
```
torch-1.4.0-pip:
  ubuntu:
    pip:
      packages: [torch==1.4.0]
```

This syntax actually works partially without this commit: `rosdep
install` installs the package with the specified version, but
incorrectly fails to detect that it was installed.

Thus, this commit relaxes the logic for determining if a package is
installed or not.

One important caveat here: there are no checks to make sure that
_that specific version_ was installed. So if the user would later
install some other version of the package in question, things like
`rosdep check` would not consider that an error.